# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ACME Salary Management",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Health Check",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/employees": {
      "get": {
        "tags": [
          "employees"
        ],
        "summary": "List Employees",
        "operationId": "list_employees_employees_get",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Max items to return (max 100)",
              "default": 20,
              "title": "Limit"
            },
            "description": "Max items to return (max 100)"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search by name or email",
              "title": "Search"
            },
            "description": "Search by name or email"
          },
          {
            "name": "department_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by department ID",
              "title": "Department Id"
            },
            "description": "Filter by department ID"
          },
          {
            "name": "country_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by country ID",
              "title": "Country Id"
            },
            "description": "Filter by country ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_EmployeeRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "employees"
        ],
        "summary": "Create Employee",
        "operationId": "create_employee_employees_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/employees/{employee_id}": {
      "get": {
        "tags": [
          "employees"
        ],
        "summary": "Get Employee",
        "operationId": "get_employee_employees__employee_id__get",
        "parameters": [
          {
            "name": "employee_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Employee Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDetailRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "employees"
        ],
        "summary": "Update Employee",
        "operationId": "update_employee_employees__employee_id__patch",
        "parameters": [
          {
            "name": "employee_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Employee Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "employees"
        ],
        "summary": "Delete Employee",
        "operationId": "delete_employee_employees__employee_id__delete",
        "parameters": [
          {
            "name": "employee_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Employee Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/employees/{employee_id}/salaries": {
      "post": {
        "tags": [
          "employees"
        ],
        "summary": "Add Salary Adjustment",
        "operationId": "add_salary_adjustment_employees__employee_id__salaries_post",
        "parameters": [
          {
            "name": "employee_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Employee Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SalaryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDetailRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/stats": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Dashboard Stats",
        "operationId": "get_dashboard_stats_dashboard_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardStats"
                }
              }
            }
          }
        }
      }
    },
    "/exchange-rates": {
      "post": {
        "tags": [
          "exchange-rates"
        ],
        "summary": "Create Exchange Rate",
        "operationId": "create_exchange_rate_exchange_rates_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExchangeRateCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExchangeRateRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/departments": {
      "get": {
        "tags": [
          "reference"
        ],
        "summary": "Get Departments",
        "operationId": "get_departments_departments_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DepartmentRead"
                  },
                  "type": "array",
                  "title": "Response Get Departments Departments Get"
                }
              }
            }
          }
        }
      }
    },
    "/currencies": {
      "get": {
        "tags": [
          "reference"
        ],
        "summary": "Get Currencies",
        "operationId": "get_currencies_currencies_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CurrencyRead"
                  },
                  "type": "array",
                  "title": "Response Get Currencies Currencies Get"
                }
              }
            }
          }
        }
      }
    },
    "/countries": {
      "get": {
        "tags": [
          "reference"
        ],
        "summary": "Get Countries",
        "operationId": "get_countries_countries_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CountryRead"
                  },
                  "type": "array",
                  "title": "Response Get Countries Countries Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ComponentTotals": {
        "properties": {
          "base_salary_usd_minor_units": {
            "type": "integer",
            "title": "Base Salary Usd Minor Units"
          },
          "housing_allowance_usd_minor_units": {
            "type": "integer",
            "title": "Housing Allowance Usd Minor Units"
          },
          "equity_usd_minor_units": {
            "type": "integer",
            "title": "Equity Usd Minor Units"
          },
          "other_allowance_usd_minor_units": {
            "type": "integer",
            "title": "Other Allowance Usd Minor Units"
          }
        },
        "type": "object",
        "required": [
          "base_salary_usd_minor_units",
          "housing_allowance_usd_minor_units",
          "equity_usd_minor_units",
          "other_allowance_usd_minor_units"
        ],
        "title": "ComponentTotals"
      },
      "CountryRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "default_currency": {
            "$ref": "#/components/schemas/CurrencyRead"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "name",
          "default_currency"
        ],
        "title": "CountryRead"
      },
      "CountryTotal": {
        "properties": {
          "country": {
            "type": "string",
            "title": "Country"
          },
          "total_salary_usd_minor_units": {
            "type": "integer",
            "title": "Total Salary Usd Minor Units"
          }
        },
        "type": "object",
        "required": [
          "country",
          "total_salary_usd_minor_units"
        ],
        "title": "CountryTotal"
      },
      "CurrencyRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "name"
        ],
        "title": "CurrencyRead"
      },
      "CurrentSalary": {
        "properties": {
          "base_salary_minor_units": {
            "type": "integer",
            "title": "Base Salary Minor Units"
          },
          "housing_allowance_minor_units": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Housing Allowance Minor Units"
          },
          "equity_minor_units": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equity Minor Units"
          },
          "other_allowance_minor_units": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Other Allowance Minor Units"
          },
          "currency": {
            "$ref": "#/components/schemas/CurrencyRead"
          },
          "valid_from": {
            "type": "string",
            "format": "date",
            "title": "Valid From"
          },
          "salary_minor_units": {
            "type": "integer",
            "title": "Salary Minor Units",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "base_salary_minor_units",
          "currency",
          "valid_from",
          "salary_minor_units"
        ],
        "title": "CurrentSalary"
      },
      "DashboardStats": {
        "properties": {
          "department_averages": {
            "items": {
              "$ref": "#/components/schemas/DepartmentAverage"
            },
            "type": "array",
            "title": "Department Averages"
          },
          "country_totals": {
            "items": {
              "$ref": "#/components/schemas/CountryTotal"
            },
            "type": "array",
            "title": "Country Totals"
          },
          "component_totals": {
            "$ref": "#/components/schemas/ComponentTotals"
          },
          "salary_distribution": {
            "items": {
              "$ref": "#/components/schemas/EmployeeSalaryPoint"
            },
            "type": "array",
            "title": "Salary Distribution"
          }
        },
        "type": "object",
        "required": [
          "department_averages",
          "country_totals",
          "component_totals",
          "salary_distribution"
        ],
        "title": "DashboardStats"
      },
      "DepartmentAverage": {
        "properties": {
          "department": {
            "type": "string",
            "title": "Department"
          },
          "average_salary_usd_minor_units": {
            "type": "integer",
            "title": "Average Salary Usd Minor Units"
          }
        },
        "type": "object",
        "required": [
          "department",
          "average_salary_usd_minor_units"
        ],
        "title": "DepartmentAverage"
      },
      "DepartmentRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "DepartmentRead"
      },
      "EmployeeCreate": {
        "properties": {
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "department_id": {
            "type": "integer",
            "title": "Department Id"
          },
          "country_id": {
            "type": "integer",
            "title": "Country Id"
          }
        },
        "type": "object",
        "required": [
          "first_name",
          "last_name",
          "email",
          "department_id",
          "country_id"
        ],
        "title": "EmployeeCreate"
      },
      "EmployeeDetailRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "department": {
            "$ref": "#/components/schemas/DepartmentRead"
          },
          "country": {
            "$ref": "#/components/schemas/CountryRead"
          },
          "current_salary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CurrentSalary"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "salary_history": {
            "items": {
              "$ref": "#/components/schemas/SalaryHistoryItem"
            },
            "type": "array",
            "title": "Salary History"
          }
        },
        "type": "object",
        "required": [
          "id",
          "first_name",
          "last_name",
          "email",
          "department",
          "country",
          "created_at",
          "updated_at",
          "salary_history"
        ],
        "title": "EmployeeDetailRead"
      },
      "EmployeeRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "department": {
            "$ref": "#/components/schemas/DepartmentRead"
          },
          "country": {
            "$ref": "#/components/schemas/CountryRead"
          },
          "current_salary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CurrentSalary"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "first_name",
          "last_name",
          "email",
          "department",
          "country",
          "created_at",
          "updated_at"
        ],
        "title": "EmployeeRead"
      },
      "EmployeeSalaryPoint": {
        "properties": {
          "department": {
            "type": "string",
            "title": "Department"
          },
          "salary_usd_minor_units": {
            "type": "integer",
            "title": "Salary Usd Minor Units"
          }
        },
        "type": "object",
        "required": [
          "department",
          "salary_usd_minor_units"
        ],
        "title": "EmployeeSalaryPoint"
      },
      "EmployeeUpdate": {
        "properties": {
          "first_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name"
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "department_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department Id"
          },
          "country_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Id"
          }
        },
        "type": "object",
        "title": "EmployeeUpdate"
      },
      "ExchangeRateCreate": {
        "properties": {
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "rate": {
            "type": "number",
            "title": "Rate"
          }
        },
        "type": "object",
        "required": [
          "currency",
          "rate"
        ],
        "title": "ExchangeRateCreate"
      },
      "ExchangeRateRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "rate": {
            "type": "number",
            "title": "Rate"
          }
        },
        "type": "object",
        "required": [
          "id",
          "currency",
          "rate"
        ],
        "title": "ExchangeRateRead"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "HealthResponse"
      },
      "PaginatedResponse_EmployeeRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EmployeeRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "PaginatedResponse[EmployeeRead]"
      },
      "SalaryCreate": {
        "properties": {
          "base_salary_minor_units": {
            "type": "integer",
            "title": "Base Salary Minor Units"
          },
          "housing_allowance_minor_units": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Housing Allowance Minor Units"
          },
          "equity_minor_units": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equity Minor Units"
          },
          "other_allowance_minor_units": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Other Allowance Minor Units"
          },
          "currency_id": {
            "type": "integer",
            "title": "Currency Id"
          },
          "valid_from": {
            "type": "string",
            "format": "date",
            "title": "Valid From"
          }
        },
        "type": "object",
        "required": [
          "base_salary_minor_units",
          "currency_id",
          "valid_from"
        ],
        "title": "SalaryCreate"
      },
      "SalaryHistoryItem": {
        "properties": {
          "base_salary_minor_units": {
            "type": "integer",
            "title": "Base Salary Minor Units"
          },
          "housing_allowance_minor_units": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Housing Allowance Minor Units"
          },
          "equity_minor_units": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equity Minor Units"
          },
          "other_allowance_minor_units": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Other Allowance Minor Units"
          },
          "currency": {
            "$ref": "#/components/schemas/CurrencyRead"
          },
          "valid_from": {
            "type": "string",
            "format": "date",
            "title": "Valid From"
          },
          "valid_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid To"
          },
          "salary_minor_units": {
            "type": "integer",
            "title": "Salary Minor Units",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "base_salary_minor_units",
          "currency",
          "valid_from",
          "valid_to",
          "salary_minor_units"
        ],
        "title": "SalaryHistoryItem"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}